New Environment "ObstacleScene1" in zKrisSandbox folder
The scene features ramps and wall-running sequences to test on, as well
as a giant "bowl" to constantly circle around. The bowl also acts as a
target to test the ramps with.

Currently it is difficult to maneuver around this thing well. The player
needs to be more forced to the ground, and be able to make sharper
turns.